pyproject.toml
[project]
name = "pi-rpc-server"
version = "0.1.0"
description = "FastAPI server for Pi coding agent RPC interface"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0,<0.116.0",
    "uvicorn[standard]>=0.32.0,<0.33.0",
    "python-multipart>=0.0.12,<0.0.13",
    "websockets>=14.0,<15.0",
    "pydantic>=2.9.0,<2.10.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "aiofiles>=24.1.0,<25.0.0",
    "markdown>=3.7.0,<4.0.0",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0,<8.4.0",
    "pytest-asyncio>=0.25.0,<0.26.0",
    "httpx>=0.28.0,<0.29.0",
    "pytest-cov>=5.0.0",
    "mypy>=1.13.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
